Tests for reference class, create constants module, fix bugs caught by tests
