+++ pydbml/__init__.py
@@ -2,10 +2,10 @@ import doctest
 import unittest
 
 from . import classes
-from pydbml.parser import PyDBML
-from pydbml.parser.blueprints import MANY_TO_ONE
-from pydbml.parser.blueprints import ONE_TO_MANY
-from pydbml.parser.blueprints import ONE_TO_ONE
+
+from pydbml.constants import MANY_TO_ONE
+from pydbml.constants import ONE_TO_MANY
+from pydbml.constants import ONE_TO_ONE
 
 
 def load_tests(loader, tests, ignore):
