+++ pydbml/constants.py
@@ -1,0 +1,3 @@
+ONE_TO_MANY = '<'
+MANY_TO_ONE = '>'
+ONE_TO_ONE = '-'
