+++ test/_test_classes.py
@@ -18,34 +18,6 @@ from pydbml.exceptions import ColumnNotFoundError
 from pydbml.exceptions import DuplicateReferenceError
 
 
-class TestDBMLObject(TestCase):
-    def test_check_attributes_for_sql(self) -> None:
-        o = SQLOjbect()
-        o.a1 = None
-        o.b1 = None
-        o.c1 = None
-        o.required_attributes = ('a1', 'b1')
-        with self.assertRaises(AttributeMissingError):
-            o.check_attributes_for_sql()
-        o.a1 = 1
-        with self.assertRaises(AttributeMissingError):
-            o.check_attributes_for_sql()
-        o.b1 = 'a2'
-        o.check_attributes_for_sql()
-
-    def test_comparison(self) -> None:
-        o1 = SQLOjbect()
-        o1.a1 = None
-        o1.b1 = 'c'
-        o1.c1 = 123
-        o2 = SQLOjbect()
-        o2.a1 = None
-        o2.b1 = 'c'
-        o2.c1 = 123
-        self.assertTrue(o1 == o2)
-        o1.a2 = True
-        self.assertFalse(o1 == o2)
-
 
 # class TestReferenceBlueprint(TestCase):
 #     def test_basic_sql(self) -> None:
@@ -101,238 +73,6 @@ class TestDBMLObject(TestCase):
 #                            on_update='no action')
 #         expected = 'CONSTRAINT "refname" FOREIGN KEY ("order_id") REFERENCES "orders ("id") ON UPDATE NO ACTION ON DELETE SET NULL'
 #         self.assertEqual(r.sql, expected)
-
-
-class TestColumn(TestCase):
-    def test_basic_sql(self) -> None:
-        r = Column(name='id',
-                   type_='integer')
-        expected = '"id" integer'
-        self.assertEqual(r.sql, expected)
-
-    def test_pk_autoinc(self) -> None:
-        r = Column(name='id',
-                   type_='integer',
-                   pk=True,
-                   autoinc=True)
-        expected = '"id" integer PRIMARY KEY AUTOINCREMENT'
-        self.assertEqual(r.sql, expected)
-
-    def test_unique_not_null(self) -> None:
-        r = Column(name='id',
-                   type_='integer',
-                   unique=True,
-                   not_null=True)
-        expected = '"id" integer UNIQUE NOT NULL'
-        self.assertEqual(r.sql, expected)
-
-    def test_default(self) -> None:
-        r = Column(name='order',
-                   type_='integer',
-                   default=0)
-        expected = '"order" integer DEFAULT 0'
-        self.assertEqual(r.sql, expected)
-
-    def test_comment(self) -> None:
-        r = Column(name='id',
-                   type_='integer',
-                   unique=True,
-                   not_null=True,
-                   comment="Column comment")
-        expected = \
-'''-- Column comment
-"id" integer UNIQUE NOT NULL'''
-        self.assertEqual(r.sql, expected)
-
-    def test_table_setter(self) -> None:
-        r1 = ReferenceBlueprint(
-            ReferenceBlueprint.MANY_TO_ONE,
-            name='refname',
-            table1='bookings',
-            col1='order_id',
-            table2='orders',
-            col2='order',
-        )
-        r2 = ReferenceBlueprint(
-            ReferenceBlueprint.MANY_TO_ONE,
-            name='refname',
-            table1='purchases',
-            col1='order_id',
-            table2='orders',
-            col2='order',
-        )
-        c = Column(
-            name='order',
-            type_='integer',
-            default=0,
-            ref_blueprints=[r1, r2]
-        )
-        t = Table('orders')
-        c.table = t
-        self.assertEqual(c.table, t)
-        self.assertEqual(c.ref_blueprints[0].table1, t.name)
-        self.assertEqual(c.ref_blueprints[1].table1, t.name)
-
-    def test_dbml_simple(self):
-        c = Column(
-            name='order',
-            type_='integer'
-        )
-        expected = '"order" integer'
-
-        self.assertEqual(c.dbml, expected)
-
-    def test_dbml_full(self):
-        c = Column(
-            name='order',
-            type_='integer',
-            unique=True,
-            not_null=True,
-            pk=True,
-            autoinc=True,
-            default='Def_value',
-            note='Note on the column',
-            comment='Comment on the column'
-        )
-        expected = \
-'''// Comment on the column
-"order" integer [pk, increment, default: 'Def_value', unique, not null, note: 'Note on the column']'''
-
-        self.assertEqual(c.dbml, expected)
-
-    def test_dbml_multiline_note(self):
-        c = Column(
-            name='order',
-            type_='integer',
-            not_null=True,
-            note='Note on the column\nmultiline',
-            comment='Comment on the column'
-        )
-        expected = \
-"""// Comment on the column
-"order" integer [not null, note: '''Note on the column
-multiline''']"""
-
-        self.assertEqual(c.dbml, expected)
-
-    def test_dbml_default(self):
-        c = Column(
-            name='order',
-            type_='integer',
-            default='String value'
-        )
-        expected = "\"order\" integer [default: 'String value']"
-        self.assertEqual(c.dbml, expected)
-
-        c.default = 3
-        expected = '"order" integer [default: 3]'
-        self.assertEqual(c.dbml, expected)
-
-        c.default = 3.33
-        expected = '"order" integer [default: 3.33]'
-        self.assertEqual(c.dbml, expected)
-
-        c.default = "(now() - interval '5 days')"
-        expected = "\"order\" integer [default: `now() - interval '5 days'`]"
-        self.assertEqual(c.dbml, expected)
-
-        c.default = 'NULL'
-        expected = '"order" integer [default: null]'
-        self.assertEqual(c.dbml, expected)
-
-        c.default = 'TRue'
-        expected = '"order" integer [default: true]'
-        self.assertEqual(c.dbml, expected)
-
-        c.default = 'false'
-        expected = '"order" integer [default: false]'
-        self.assertEqual(c.dbml, expected)
-
-class TestIndex(TestCase):
-    def test_basic_sql(self) -> None:
-        t = Table('products')
-        t.add_column(Column('id', 'integer'))
-        r = Index(subject_names=['id'],
-                  table=t)
-        t.add_index(r)
-        expected = 'CREATE INDEX ON "products" ("id");'
-        self.assertEqual(r.sql, expected)
-
-    def test_comment(self) -> None:
-        t = Table('products')
-        t.add_column(Column('id', 'integer'))
-        r = Index(subject_names=['id'],
-                  table=t,
-                  comment='Index comment')
-        t.add_index(r)
-        expected = \
-'''-- Index comment
-CREATE INDEX ON "products" ("id");'''
-
-        self.assertEqual(r.sql, expected)
-
-    def test_unique_type_composite(self) -> None:
-        t = Table('products')
-        t.add_column(Column('id', 'integer'))
-        t.add_column(Column('name', 'varchar'))
-        r = Index(subject_names=['id', 'name'],
-                  table=t,
-                  type_='hash',
-                  unique=True)
-        t.add_index(r)
-        expected = 'CREATE UNIQUE INDEX ON "products" USING HASH ("id", "name");'
-        self.assertEqual(r.sql, expected)
-
-    def test_pk(self) -> None:
-        t = Table('products')
-        t.add_column(Column('id', 'integer'))
-        t.add_column(Column('name', 'varchar'))
-        r = Index(subject_names=['id', 'name'],
-                  table=t,
-                  pk=True)
-        t.add_index(r)
-        expected = 'PRIMARY KEY ("id", "name")'
-        self.assertEqual(r.sql, expected)
-
-    def test_composite_with_expression(self) -> None:
-        t = Table('products')
-        t.add_column(Column('id', 'integer'))
-        r = Index(subject_names=['id', '(id*3)'],
-                  table=t)
-        t.add_index(r)
-        self.assertEqual(r.subjects, [t['id'], '(id*3)'])
-        expected = 'CREATE INDEX ON "products" ("id", (id*3));'
-        self.assertEqual(r.sql, expected)
-
-    def test_dbml_simple(self):
-        i = Index(
-            ['id']
-        )
-
-        expected = 'id'
-        self.assertEqual(i.dbml, expected)
-
-    def test_dbml_composite(self):
-        i = Index(
-            ['id', '(id*3)']
-        )
-
-        expected = '(id, `id*3`)'
-        self.assertEqual(i.dbml, expected)
-
-    def test_dbml_full(self):
-        i = Index(
-            ['id', '(getdate())'],
-            name='Dated id',
-            unique=True,
-            type_='hash',
-            pk=True,
-            note='Note on the column',
-            comment='Comment on the index'
-        )
-        expected = \
-'''// Comment on the index
-(id, `getdate()`) [name: 'Dated id', pk, unique, type: hash, note: 'Note on the column']'''
 
 
 class TestTable(TestCase):
@@ -609,126 +349,6 @@ class TestTableReference(TestCase):
         expected = 'CONSTRAINT "country_name" FOREIGN KEY ("name", "country") REFERENCES "names" ("name_val", "country_val") ON UPDATE CASCADE ON DELETE SET NULL'
         self.assertEqual(ref.sql, expected)
 
-class TestReference(TestCase):
-    def test_sql_single(self):
-        t = Table('products')
-        c1 = Column('name', 'varchar2')
-        t.add_column(c1)
-        t2 = Table('names')
-        c2 = Column('name_val', 'varchar2')
-        t2.add_column(c2)
-        ref = Reference('>', t, c1, t2, c2)
-
-        expected = 'ALTER TABLE "products" ADD FOREIGN KEY ("name") REFERENCES "names" ("name_val");'
-        self.assertEqual(ref.sql, expected)
-
-    def test_sql_reverse(self):
-        t = Table('products')
-        c1 = Column('name', 'varchar2')
-        t.add_column(c1)
-        t2 = Table('names')
-        c2 = Column('name_val', 'varchar2')
-        t2.add_column(c2)
-        ref = Reference('<', t, c1, t2, c2)
-
-        expected = 'ALTER TABLE "names" ADD FOREIGN KEY ("name_val") REFERENCES "products" ("name");'
-        self.assertEqual(ref.sql, expected)
-
-    def test_sql_multiple(self):
-        t = Table('products')
-        c11 = Column('name', 'varchar2')
-        c12 = Column('country', 'varchar2')
-        t.add_column(c11)
-        t.add_column(c12)
-        t2 = Table('names')
-        c21 = Column('name_val', 'varchar2')
-        c22 = Column('country_val', 'varchar2')
-        t2.add_column(c21)
-        t2.add_column(c22)
-        ref = Reference('>', t, [c11, c12], t2, (c21, c22))
-
-        expected = 'ALTER TABLE "products" ADD FOREIGN KEY ("name", "country") REFERENCES "names" ("name_val", "country_val");'
-        self.assertEqual(ref.sql, expected)
-
-    def test_sql_full(self):
-        t = Table('products')
-        c11 = Column('name', 'varchar2')
-        c12 = Column('country', 'varchar2')
-        t.add_column(c11)
-        t.add_column(c12)
-        t2 = Table('names')
-        c21 = Column('name_val', 'varchar2')
-        c22 = Column('country_val', 'varchar2')
-        t2.add_column(c21)
-        t2.add_column(c22)
-        ref = Reference(
-            '>',
-            t,
-            [c11, c12],
-            t2,
-            (c21, c22),
-            name="country_name",
-            comment="Multiline\ncomment for the constraint",
-            on_update="CASCADE",
-            on_delete="SET NULL"
-        )
-
-        expected = \
-'''-- Multiline
--- comment for the constraint
-ALTER TABLE "products" ADD CONSTRAINT "country_name" FOREIGN KEY ("name", "country") REFERENCES "names" ("name_val", "country_val") ON UPDATE CASCADE ON DELETE SET NULL;'''
-
-        self.assertEqual(ref.sql, expected)
-
-    def test_dbml_simple(self):
-        t = Table('products')
-        c1 = Column('id', 'integer')
-        c2 = Column('name', 'varchar2')
-        t.add_column(c1)
-        t.add_column(c2)
-        t2 = Table('names')
-        c21 = Column('name_val', 'varchar2')
-        t2.add_column(c21)
-        ref = Reference('>', t, c2, t2, c21)
-
-        expected = \
-'''Ref {
-    "products"."name" > "names"."name_val"
-}'''
-        self.assertEqual(ref.dbml, expected)
-
-    def test_dbml_full(self):
-        t = Table('products')
-        c1 = Column('id', 'integer')
-        c2 = Column('name', 'varchar2')
-        c3 = Column('country', 'varchar2')
-        t.add_column(c1)
-        t.add_column(c2)
-        t.add_column(c3)
-        t2 = Table('names')
-        c21 = Column('name_val', 'varchar2')
-        c22 = Column('country', 'varchar2')
-        t2.add_column(c21)
-        t2.add_column(c22)
-        ref = Reference(
-            '<',
-            t,
-            [c2, c3],
-            t2,
-            (c21, c22),
-            name='nameref',
-            comment='Reference comment\nmultiline',
-            on_update='CASCADE',
-            on_delete='SET NULL'
-        )
-
-        expected = \
-'''// Reference comment
-// multiline
-Ref nameref {
-    "products".("name", "country") < "names".("name_val", "country") [update: CASCADE, delete: SET NULL]
-}'''
-        self.assertEqual(ref.dbml, expected)
 
 
 class TestNote(TestCase):
